Migrate the message dispatcher and the events handler from `Gaia` to `Engine`
- It made little sense to set up `Engine` and the events handler and `Events` in `Gaia` and then pass references to each other as "siblings" while in practice `Events` is the child of `Engine`
